Fix import in differ
